Performance boost by using ETS tables for buffering messages.
Thanks to much advice from the [Erlang Forum](
  https://erlangforums.com/t/efficient-storage-using-lists/4988
), now using ETS table and binary representations for messages objects.

The problem was that using lists to buffer message instances in hash
representation was sub-optimal when dealing with many thousands of messages.

Tests:

- https://flows.red-erik.org/f/9a1bdfdcd7b2be8e
- https://flows.red-erik.org/f/866410b56fa42447

This commit replaces the approach of using lists and hashes with ETS tables
per process and converting messages to a binary representation.

There are probably many more optimisations possible but these have lead to
the Erlang-Red implementation being faster than Node-RED when executing the
test flows. So that's a good step in the right direction.