[NEW] CaptureIO Node
Node for implementing the I/O protocol and obtaining output from
processes and feeding that output into Erlang-Red as messages.

This allows using io:format(...) inside function nodes and having that
appear in the debug sidebar inside Erlang-Red.

At the moment there is too much appearing in the console window that should
be streamed into Erlang-Red. This approach takes the path of having an extra
node that can be configured to capture I/O from specific nodes instead
of a blanket solution that captures all I/O.

This has the advantage of better understand what generated the I/O and being
about to react to that.

The implemntation is beta with kinks to be ironed out;

- one or many processes for implementing the I/O protocol. ATM each new function
  processes gets a new group leader process and both die when the function
  is done. This does not have to be the case, the group leader process can
  remain and can be reused for other processes. This will require more logic
  but less spinning up of processes.

- there is a implicit ordering of process and configuring a capture for an
  existing process requires the capture to be defined before the process
  is embedded into Erlang-Red. Confused? Yes, so am I. There should never
  be an order in which nodes are defined - implicit dependencies are not
  good in flow based programming.